Added "Become creator" functionality

+++ SharedGoals/Controllers/GoalsController.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using SharedGoals.Data;
 using SharedGoals.Data.Models;
+using SharedGoals.Infrastructure;
 using SharedGoals.Models.Goals;
 using System;
 using System.Collections.Generic;
@@ -31,6 +33,7 @@ namespace SharedGoals.Controllers
             return View(goals);
         }
 
+        [Authorize]
         public IActionResult Details(int id)
         {
             var goal = this.dbContext.Goals.FirstOrDefault(g => g.Id == id);
@@ -54,14 +57,35 @@ namespace SharedGoals.Controllers
             return View(goalData);
         }
 
-        public IActionResult Create() => View(new CreateGoalFormModel
+        [Authorize]
+        public IActionResult Create()
         {
-            Tags = GetGoalTags()
-        });
+            if (!this.UserIsCreator())
+            {
+                return RedirectToAction(nameof(CreatorsController.Become), "Creators");
+            }
+
+            return View(new CreateGoalFormModel
+            {
+                Tags = GetGoalTags()
+            });
+        }
 
         [HttpPost]
+        [Authorize]
         public IActionResult Create(CreateGoalFormModel goal)
         {
+            var creatorId = this.dbContext
+                .Creators
+                .Where(c => c.UserId == this.User.GetId())
+                .FirstOrDefault()
+                .Id;
+
+            if (creatorId == null)
+            {
+                return RedirectToAction(nameof(CreatorsController.Become), "Creators");
+            }
+
             if (!this.dbContext.Tags.Any(c => c.Id == goal.TagId))
             {
                 this.ModelState.AddModelError(nameof(goal.TagId), "Tag does not exist.");
@@ -81,26 +105,39 @@ namespace SharedGoals.Controllers
 
             var goalData = new Goal
             {
-               Name = goal.Name,
-               Description = goal.Description,
-               CreatedOn = DateTime.UtcNow,
-               DueDate = goal.DueDate,
-               TagId = goal.TagId
+                Name = goal.Name,
+                Description = goal.Description,
+                CreatedOn = DateTime.UtcNow,
+                DueDate = goal.DueDate,
+                TagId = goal.TagId,
+                CreatorId = creatorId
             };
 
             this.dbContext.Goals.Add(goalData);
             this.dbContext.SaveChanges();
 
-            return Redirect("/");
+            return RedirectToAction("All", "Goals");
         }
 
+        [Authorize]
         public IActionResult Delete(int id)
         {
             var goal = this.dbContext.Goals.FirstOrDefault(g => g.Id == id);
 
+            if (!this.UserIsCreator())
+            {
+                return RedirectToAction(nameof(CreatorsController.Become), "Creators");
+            }
+
             if (goal == null)
             {
-                return this.View();
+                return View();
+            }
+
+            var currentUser = this.dbContext.Creators.FirstOrDefault(u => u.UserId == this.User.GetId());
+            if (goal.CreatorId != currentUser.Id)
+            {
+                return Unauthorized("You cannot delete a goal of another creator!");
             }
 
             var tagName = this.dbContext.Tags.FirstOrDefault(t => t.Id == goal.TagId).Name;
@@ -117,6 +154,7 @@ namespace SharedGoals.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public IActionResult Delete(GoalListingViewModel goalModel)
         {
             var goal = this.dbContext.Goals.FirstOrDefault(g => g.Id == goalModel.Id);
@@ -125,11 +163,23 @@ namespace SharedGoals.Controllers
                 return this.View();
             }
 
+            if (!this.UserIsCreator())
+            {
+                return RedirectToAction(nameof(CreatorsController.Become), "Creators");
+            }
+
+            var currentUser = this.dbContext.Creators.FirstOrDefault(u => u.UserId == this.User.GetId());
+            if (goal.CreatorId != currentUser.Id)
+            {
+                return Unauthorized("You cannot delete a goal of another creator!");
+            }
+
             dbContext.Goals.Remove(goal);
             dbContext.SaveChanges();
             return this.RedirectToAction("All");
         }
 
+        [Authorize]
         public IActionResult Edit(int id)
         {
             var goal = this.dbContext.Goals.FirstOrDefault(g => g.Id == id);
@@ -137,6 +187,17 @@ namespace SharedGoals.Controllers
             if (goal == null)
             {
                 return this.View();
+            }
+
+            if (!this.UserIsCreator())
+            {
+                return RedirectToAction(nameof(CreatorsController.Become), "Creators");
+            }
+
+            var currentUser = this.dbContext.Creators.FirstOrDefault(u => u.UserId == this.User.GetId());
+            if (goal.CreatorId != currentUser.Id)
+            {
+                return Unauthorized("You cannot edit a goal of another creator!");
             }
 
             CreateGoalFormModel model = new CreateGoalFormModel()
@@ -160,6 +221,17 @@ namespace SharedGoals.Controllers
                 return this.View();
             }
 
+            if (!this.UserIsCreator())
+            {
+                return RedirectToAction(nameof(CreatorsController.Become), "Creators");
+            }
+
+            var currentUser = this.dbContext.Creators.FirstOrDefault(u => u.UserId == this.User.GetId());
+            if (goal.CreatorId != currentUser.Id)
+            {
+                return Unauthorized("You cannot edit a goal of another creator!");
+            }
+
             if (!this.dbContext.Tags.Any(c => c.Id == bindingModel.TagId))
             {
                 this.ModelState.AddModelError(nameof(bindingModel.TagId), "Tag does not exist.");
@@ -173,7 +245,7 @@ namespace SharedGoals.Controllers
             if (!this.ModelState.IsValid)
             {
                 bindingModel.Tags = GetGoalTags();
-                return this.View(bindingModel); 
+                return this.View(bindingModel);
             }
 
             goal.Name = bindingModel.Name;
@@ -184,6 +256,11 @@ namespace SharedGoals.Controllers
             dbContext.SaveChanges();
             return this.RedirectToAction("All");
         }
+
+        private bool UserIsCreator()
+            => this.dbContext
+                .Creators
+                .Any(c => c.UserId == this.User.GetId());
 
         private IEnumerable<GoalTagViewModel> GetGoalTags()
             => this.dbContext
